Use 'devise' theme
It's got a lot of what I'm looking for. First, it's oriented to
writing blog posts. That may sound trite, but there are so. many.
themes. It took me a while to find something promising "out of the
box". I have close to zero interest in creating my own Hugo
module.

Also: it's extremely minimal, and it doesn't preview posts, which
in my experience makes a mess of syntax highlighting.